+++ src/Symfony/Component/Form/CHANGELOG.md
@@ -7,6 +7,9 @@ CHANGELOG
  * added `Symfony\Component\Form\FormErrorIterator::findByCodes()`
  * added `getTypedExtensions`, `getTypes`, and `getTypeGuessers` to `Symfony\Component\Form\Test\FormIntegrationTestCase`
  * added `FormPass`
+ * deprecated `choice_attr` option as array of nested arrays mapped by indexes
+ * deprecated `choice_attr` option as string or `PropertyPath` instance
+ * added `choice_label_attr` option to `ChoiceType` to use when expanded
 
 3.2.0
 -----